try to set up route
